fix: author sharing image due to Hugo breaking change
When sharing the page of an author profile (/author/author-name), the avatar image was no longer shown as the sharing image.

Hugo team made a breaking change in recent Hugo versions to rename the page `.Kind` to `term` instead of `taxonomy` in this case so the statement was no longer triggered.

Fix #1987