feat: show button links in listings irrespective of Type
For example, if a blog post has an attachment, it will now be linked in the listing as well as in its page header.

May also reduce excess spacing by preventing empty attachment DIV in the case of no attachments/buttons since it now checks if a page has buttons.