fix: Language name doesn't appear for zh-Hant
Language name doesn't appear when LanguageCode has a sub-code

The issue appears to be an inconsistency/bug with Hugo's `.Translations` lowercasing zh-Hant to zh-hant without offering a way to get the language code as it was defined.

We can workaround it for now by converting to lowercase to align with Hugo's current implementation.

Fix #2256